+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.3
+- Fixed /acunlock command and unlocking through GUI
+- Only show buttons in GUI that the player has permission to
+
 ## 0.0.2
 - Configurable command aliases